NX_Femtetのエラーハンドリングを完成

+++ src/PyFemtet/opt/_NX_Femtet.py
@@ -8,8 +8,22 @@ from win32com.client import Dispatch, DispatchEx
 import win32process
 import win32api
 import win32con
+from pywinauto import Application
+from pywinauto.application import ProcessNotFoundError
 
-import numpy as np
+import threading
+
+from PyFemtet.opt.core import FEMSystem
+
+
+
+# import re
+import warnings
+pattern = r"32-bit"
+warnings.filterwarnings("ignore", category=UserWarning, message=pattern)
+# warnings.filterwarnings("ignore", category=UserWarning, module='pywinauto')
+# warnings.filterwarnings('ignore', category=UserWarning)
+
 
 here, me = os.path.split(__file__)
 
@@ -27,8 +41,9 @@ PATH_REF = r'C:\Program Files\Femtet_Ver2023.1.0_64bit\Program\Macro32\FemtetRef
 PATH_JOURNAL = os.path.abspath(os.path.join(here, 'update_model_parameter.py'))
 
 
-
-def _f(functions, arguments): # インスタンスメソッドにしたら動かない クラスメソッドにしても動かない。なんでだろう？
+def _f(functions, arguments):
+    # インスタンスメソッドにしたら動かない
+    # クラスメソッドにしても動かない。なんでだろう？
     # 新しいプロセスで呼ぶ関数。
     # 新しい Femtet を作って objectives を計算する
     # その後、プロセスは死ぬので Femtet は解放される
@@ -42,13 +57,13 @@ def _f(functions, arguments): # インスタンスメソッドにしたら動か
         ret.append(func(Femtet, *args, **kwargs))
     return ret
 
-from PyFemtet.opt.core import FEMSystem
 
 class NX_Femtet(FEMSystem):
     def __init__(self, path_prt):
         self._path_prt = os.path.abspath(path_prt)
         self._path_bas = None
         self._path_xlsm = None
+        self._stop_excel_watcher = False
     
     def set_bas(self, path_bas):
         self._path_bas = os.path.abspath(path_bas)
@@ -68,16 +83,19 @@ class NX_Femtet(FEMSystem):
         try:
             self._update_model(df)
         except:
+            print('failed to create model via NX')
             raise ModelError
 
         try:
             self._setup_new_Femtet()
         except:
-            raise ModelError
+            print('failed to setup model via Excel-Femtet')
+            raise ModelError # TODO: Excel のエラーを検出する
 
         try:
             self._run_new_Femtet(objectives)
         except:
+            print('failed to mesh or solve via FEMOpt-Femtet')
             raise SolveError
  
         return self.objectiveValues
@@ -87,7 +105,8 @@ class NX_Femtet(FEMSystem):
         # prt と同じ名前の x_t ができる
         # 先にそれを消しておく
         path_x_t = os.path.splitext(self._path_prt)[0] + '.x_t'
-        os.remove(path_x_t)
+        if os.path.isfile(path_x_t):
+            os.remove(path_x_t)
         exe = r'%UGII_BASE_DIR%\NXBIN\run_journal.exe'
         tmp = dict(zip(df.name.values, df.value.values.astype(str)))
         strDict = json.dumps(tmp)
@@ -98,7 +117,7 @@ class NX_Femtet(FEMSystem):
             shell=True,
             cwd=os.path.dirname(self._path_prt))
         # この時点で x_t ファイルがなければモデルエラー
-        if os.path.isfile(path_x_t):
+        if not os.path.isfile(path_x_t):
             from PyFemtet.opt.core import ModelError
             raise ModelError
 
@@ -136,22 +155,32 @@ class NX_Femtet(FEMSystem):
             # マクロのセットアップ
             self._set_reference_of_new_excel(wb)
 
+        # TODO: この処理の検証：保存したことにする（エラーで強制終了されるときにこの子にアクセスできないからここで Saved しとけば「アイテムの回復」に出てこないことを期待、まだ未テスト）
+        wb.Saved = True
+        
+        # excel VBA エラー監視スレッドの開始
+        self._stop_excel_watcher = False
+        _, pid = win32process.GetWindowThreadProcessId(self.excel.Hwnd)
+        t = threading.Thread(target=_excel_watcher, args=(pid, self))
+        t.start()
 
         # Femtet セットアップの実行
-        self.excel.Run('FemtetMacro.FemtetMain')
-
-        # # マクロの破棄
-        # if self._path_bas is not None:
-        #     self.excel.Run('HubModule.ReleaseFemtetMacroBas')    
-        #     self.excel.Run('HubModule.ReleaseReference', 'FemtetMacro')
-        #     self.excel.Run('HubModule.ReleaseReference', 'FemtetReference')
-
-        # 保存せずに閉じる
-        wb.Saved = True
-        wb.Close()
-
-        # 終了
-        self._close_excel_by_force()
+        from .core import ModelError
+        from pythoncom import com_error
+        try:
+            self.excel.Run('FemtetMacro.FemtetMain')
+        except com_error:
+            # エラーが発生している場合は excel_watcher が
+            # excel のプロセスを終了しているから何もしなくていい
+            raise ModelError
+        else:
+            # 問題なく終了すれば、スレッドを閉じるフラグを立てる
+            self._stop_excel_watcher = True
+            sleep(1) # スレッドの終了を一応待つ
+            # 保存せずに閉じる
+            wb.Close()
+            # 終了
+            self._close_excel_by_force()
 
     def _run_new_Femtet(self, objectives):
         # 関数を適用する
@@ -176,18 +205,53 @@ class NX_Femtet(FEMSystem):
 
 
 
-# #### サンプル関数
-# from win32com.client import constants
-# def get_flow(Femtet):
-#     Gogh = Femtet.Gogh
-#     Gogh.Pascal.Vector = constants.PASCAL_VELOCITY_C
-#     _, ret = Gogh.SimpleIntegralVectorAtFace_py([2], [0], constants.PART_VEC_Y_PART_C)
-#     flow = ret.Real
-#     return flow
+def _get_excel_state(pid, timeout=1):
+    '''
+    excel がなければ 0, あれば 1
+    あって、さらに「Microsoft Visual Basic」ダイアログがあれば -1 を返す
+    '''
+    
+    # Excelのアプリケーションを取得
+    try:
+        app = Application().connect(process=pid)
+    except ProcessNotFoundError:
+        return 0
+    
+    # ダイアログの存在の有無
+    try:
+        app.window(title="Microsoft Visual Basic").wait('exists', timeout=timeout)
+        return -1
+    except:
+        sleep(timeout)
+        return 1
 
 
-# if __name__=='__main__':
-#     FEMOpt = NX_Femtet()
-#     df = None
+def _excel_watcher(pid, stopper:NX_Femtet):
+    print('excel_watcher has been launched')
+    # threading に呼ばれ、1秒ごとに excel の状態を監視
+    # エラーが出ていれば excel を終了して仕事を終了する
+    # excel が終了していれば仕事を終了する
+    while True:
+        excel_state = _get_excel_state(pid, timeout=1)
+        print('excel state is ', excel_state)
+        if excel_state==-1:
+            # excel を強制終了して break する
+            try:
+                handle = win32api.OpenProcess(win32con.PROCESS_TERMINATE, 0, pid)
+                if handle:
+                    win32api.TerminateProcess(handle, 0)
+                    win32api.CloseHandle(handle)
+            except:
+                pass
+            break
+        if excel_state==0:
+            # 終了する
+            break
+        if stopper._stop_excel_watcher:
+            # 終了する
+            break
+    print('excel_watcher will be terminated successfully')
+    return None # thread 終了
+
 #     print(FEMOpt.f(df, [get_flow]))
